[dev-1564] moving components to shared/
